% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bayesian lasso mfvb}
\alias{bayesian lasso mfvb}
\title{bayesian lasso mfvb}
\usage{
TBA
}
\value{
TBA
}
\description{
TBA
}
\details{
TBA
}
\examples{
rnorm(10)

}
\references{
Where did this come from?
}
